refactor(model): divide model.h into algebra with subfolders
BREAKING CHANGE:
Declarations of Point, Vector and Angle are now to be found in algebra/
and included through algebra.h
