+++ src/model/algebra/algebra.h
@@ -1,0 +1,8 @@
+#ifndef EVE_ALGEBRA_H
+#define EVE_ALGEBRA_H
+
+#include "angle/angle.h"
+#include "point/point.h"
+#include "vector/vector.h"
+
+#endif //EVE_ALGEBRA_H
