+++ src/model/algebra/algebra_conf.h
@@ -1,0 +1,10 @@
+#ifndef EVE_ALGEBRA_CONF_H
+#define EVE_ALGEBRA_CONF_H
+
+#define CONST_MULTIPLIER 1000
+
+struct Angle;
+struct Point;
+struct Vector;
+
+#endif //EVE_ALGEBRA_CONF_H
